conditionally render etherspot or archanova history
