+++ src/components/HistoryList/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "HistoryList",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./HistoryList.js"
+}
